Standard settings to separate module.
