+++ lib/graph3d/Graph3d.js
@@ -9,167 +9,10 @@ var Filter       = require('./Filter');
 var Slider       = require('./Slider');
 var StepNumber   = require('./StepNumber');
 var support      = require('./stylehandlers/Support');
+var Settings     = require('./Settings');
 var StyleHandler = require('./stylehandlers/StyleHandler');
 
-
-/**
- * Field names in the options hash which are of relevance to the user.
- *
- * Specifically, these are the fields which require no special handling,
- * and can be directly copied over.
- */
-var OPTIONKEYS = [
-  'width',
-  'height',
-  'xCenter',
-  'yCenter',
-  'filterLabel',
-  'legendLabel',
-  'xLabel',
-  'yLabel',
-  'zLabel',
-  'xValueLabel',
-  'yValueLabel',
-  'zValueLabel',
-  'dotSizeRatio',
-  'showGrid',
-  'showPerspective',
-  'showShadow',
-  'showAnimationControls',
-  'keepAspectRatio',
-  'verticalRatio',
-  'animationInterval',
-  'animationPreload',
-  'animationAutoStart',
-  'xBarWidth',
-  'yBarWidth',
-  'xMin',
-  'xStep',
-  'xMax',
-  'yMin',
-  'yStep',
-  'yMax',
-  'zMin',
-  'zStep',
-  'zMax',
-  'valueMin',
-  'valueMax',
-  'axisColor',
-  'gridColor',
-  'showLegend'
-];
-
-
-/**
- * Default values for certain option fields.
- *
- * These are the values used when a Graph3d instance is initialized
- * without custom settings.
- *
- * If a field is not in this list, a default value of 'undefined' can 
- * be assumed. Of course, it does no harm to set a field explicitly to
- * 'undefined' here.
- *
- * A value of 'undefined' here normally means:
- *
- *     'derive from current data and graph style'
- *
- * In the code, this is indicated by the comment 'auto by default'.
- */ 
-var DEFAULTS = {
-  width            : '400px',
-  height           : '400px',
-  xLabel           : 'x',
-  yLabel           : 'y',
-  zLabel           : 'z',
-  xValueLabel      : function(v) { return v; },
-  yValueLabel      : function(v) { return v; },
-  zValueLabel      : function(v) { return v; },
-  filterLabel      : 'time',
-  legendLabel      : 'value',
-  showPerspective  : true,
-  showGrid         : true,
-  keepAspectRatio  : true,
-  showShadow       : false,
-  verticalRatio    : 0.5,       // 0.1 to 1.0, where 1.0 results in a 'cube'
-  animationInterval: 1000,      // milliseconds
-  animationPreload : false,
-  xMin             : 0,
-  xStep            : undefined, // auto by default
-  xMax             : 1,
-  yMin             : 0,
-  yStep            : undefined, // auto by default
-  yMax             : 1,
-  zMin             : 0,
-  zStep            : undefined, // auto by default
-  zMax             : 1,
-  valueMin         : 0,
-  valueMax         : 1,
-  xBarWidth        : 1,
-  yBarWidth        : 1,
-  axisColor        : '#4D4D4D',
-  gridColor        : '#D3D3D3',
-  dotSizeRatio     : 0.02,      // size of the dots as a fraction of the graph width
-  showLegend       : undefined, // auto by default (based on graph style)
-  xBarWidth        : undefined, // auto by default
-  yBarWidth        : undefined, // auto by default
-
-
-  // Following not in defaults (yet) but present in user settings
-  //'xCenter',
-  //'yCenter',
-  //'showAnimationControls',
-  //'animationAutoStart',
-
-  //
-  // Sleight of hand:
-  //
-  // Values below intentionally start with a capital; these are NOT copied to
-  // the active options as used in the Graph3d instance.
-  // The values mentioned are used directly in the code.
-  //
-  // This hackish approach retains the original working of the code.
-  // Chances are, it will disappear with future progress.
-  //
-  XCenter       : '55%',
-  YCenter       : '50%',
-  ShowLegend    : false,
-  Style         : StyleHandler.STYLE.DOT,
-
-  CameraPosition: {
-    horizontal: 1.0,
-    vertical  : 0.5,
-    distance  : 1.7
-  },  
-
-  DataColor     : {
-    fill       : '#7DC1FF',
-    stroke     : '#3267D2',
-    strokeWidth: 1          // px
-  }
-
-  // Other values used explicitly in code.
-  // These will be 'undefined', which is OK.
-  //
-  // XMin, XMax, XStep,
-  // YMin, YMax, YStep,
-  // ZMin, ZMax, ZStep,
-  // ValueMin, ValueMax,
-}; 
-
-
-/**
- * Local helper function to tone down the length of the construct:
- *
- *         target  = (A !== undefined) ? A : B;
- */
-function select(valueA, valueB) {
-  if (valueA !== undefined) {
-    return valueA;
-  } else {
-    return valueB;
-  }
-}
+var select = support.select;     // Local alias for given function
 
 
 /**
@@ -191,20 +34,12 @@ function Graph3d(container, data, options) {
   // create variables and set default values
   this.containerElement = container;
 
-  support.copy(DEFAULTS, this, OPTIONKEYS);
-
-  // Other fields which apparently are not part of the user options
-  this.margin          = 10;        // px
-  this.showGrayBottom  = false; // TODO: this does not work correctly
-  this.showTooltip     = false;
-
-  this._setStyle(DEFAULTS.Style, this);
-  this._setCameraPosition(DEFAULTS.CameraPosition, this);
+  Settings.setDefaults(this);
 
   // TODO: set eye.z about 3/4 of the width of the window?
   this.eye = new Point3d(0, 0, -1);
 
-  this.dataTable = null;  // The original data table
+  this.dataTable  = null;  // The original data table
   this.dataPoints = null; // The table with point objects
 
   // the column indexes
@@ -215,8 +50,6 @@ function Graph3d(container, data, options) {
   this.colFilter = undefined;
 
   // TODO: customize axis range
-
-  this._setDataColor(DEFAULTS.DataColor, this);
 
   // create a frame and canvas
   this.create();
@@ -358,34 +191,6 @@ Graph3d.prototype._convertTranslationToScreen = function(translation) {
 };
 
 
-/**
- * Set the background styling for the graph
- * @param {string | {fill: string, stroke: string, strokeWidth: string}} backgroundColor
- */
-Graph3d.prototype._setBackgroundColor = function(backgroundColor) {
-  var fill = 'white';
-  var stroke = 'gray';
-  var strokeWidth = 1;
-
-  if (typeof(backgroundColor) === 'string') {
-    fill = backgroundColor;
-    stroke = 'none';
-    strokeWidth = 0;
-  } else if (typeof(backgroundColor) === 'object') {
-    fill        = select(backgroundColor.fill       , fill);
-    stroke      = select(backgroundColor.stroke     , stroke);
-    strokeWidth = select(backgroundColor.strokeWidth, strokeWidth);
-  } else if  (backgroundColor === undefined) {
-    // use defaults
-  } else {
-    throw 'Unsupported type of backgroundColor';
-  }
-
-  this.frame.style.backgroundColor = fill;
-  this.frame.style.borderColor = stroke;
-  this.frame.style.borderWidth = strokeWidth + 'px';
-  this.frame.style.borderStyle = 'solid';
-};
 
 
 /**
@@ -504,6 +309,8 @@ Graph3d.prototype._dataInitialize = function (rawData, style) {
   // For Bar graphs, take bar dimensions into account
   this._styleHandler.adjustForBarWidth(this, data, xRange, yRange);
 
+  var DEFAULTS = Settings.DEFAULTS;
+
   // calculate minimums and maximums
   this.xMin = select(DEFAULTS.XMin, xRange.min);
   this.xMax = select(DEFAULTS.XMax, xRange.max);
@@ -561,17 +368,17 @@ Graph3d.prototype.create = function () {
   this.frame.appendChild(this.frame.canvas);
 
   var noCanvas = document.createElement( 'DIV' );
-  noCanvas.style.color = 'red';
+  noCanvas.style.color      = 'red';
   noCanvas.style.fontWeight =  'bold' ;
-  noCanvas.style.padding =  '10px';
-  noCanvas.innerHTML =  'Error: your browser does not support HTML canvas';
+  noCanvas.style.padding    =  '10px';
+  noCanvas.innerHTML        =  'Error: your browser does not support HTML canvas';
   this.frame.canvas.appendChild(noCanvas);
 
   this.frame.filter = document.createElement( 'div' );
   this.frame.filter.style.position = 'absolute';
-  this.frame.filter.style.bottom = '0px';
-  this.frame.filter.style.left = '0px';
-  this.frame.filter.style.width = '100%';
+  this.frame.filter.style.bottom   = '0px';
+  this.frame.filter.style.left     = '0px';
+  this.frame.filter.style.width    = '100%';
   this.frame.appendChild(this.frame.filter);
 
   // add event listeners to handle moving and zooming the contents
@@ -582,11 +389,11 @@ Graph3d.prototype.create = function () {
   var ontooltip    = function (event) {me._onTooltip(event);};
   // TODO: these events are never cleaned up... can give a 'memory leakage'
 
-  util.addEventListener(this.frame.canvas, 'keydown', onkeydown);
-  util.addEventListener(this.frame.canvas, 'mousedown', onmousedown);
+  util.addEventListener(this.frame.canvas, 'keydown'   , onkeydown);
+  util.addEventListener(this.frame.canvas, 'mousedown' , onmousedown);
   util.addEventListener(this.frame.canvas, 'touchstart', ontouchstart);
   util.addEventListener(this.frame.canvas, 'mousewheel', onmousewheel);
-  util.addEventListener(this.frame.canvas, 'mousemove', ontooltip);
+  util.addEventListener(this.frame.canvas, 'mousemove' , ontooltip);
 
   // add the new graph to the container element
   this.containerElement.appendChild(this.frame);
@@ -666,8 +473,8 @@ Graph3d.prototype._resizeCenter = function() {
     return ret;
   };
 
-  this.xcenter = parse(DEFAULTS.XCenter);
-  this.ycenter = parse(DEFAULTS.YCenter);
+  this.xcenter = parse(Settings.DEFAULTS.XCenter);
+  this.ycenter = parse(Settings.DEFAULTS.YCenter);
 };
 
 
@@ -754,121 +561,13 @@ Graph3d.prototype.setData = function (data) {
 };
 
 
-Graph3d.prototype._setDataColor = function (dataColor, dst) {
-  if (dataColor === undefined) {
-    return;    // Nothing to do
-  }
-
-  if (dst.dataColor === undefined) {
-    dst.dataColor = {};
-  }
-
-  if (typeof dataColor === 'string') {
-    dst.dataColor.fill   = dataColor;
-    dst.dataColor.stroke = dataColor;
-  }
-  else {
-    if (dataColor.fill) {
-      dst.dataColor.fill = dataColor.fill;
-    }
-    if (dataColor.stroke) {
-      dst.dataColor.stroke = dataColor.stroke;
-    }
-    if (dataColor.strokeWidth !== undefined) {
-      dst.dataColor.strokeWidth = dataColor.strokeWidth;
-    }
-  }
-};
-
-
-Graph3d.prototype._setCameraPosition = function (cameraPosition, dst) {
-  var camPos = cameraPosition;
-  if (camPos === undefined) {
-    return;
-  }
-
-  if (dst.camera === undefined) {
-    dst.camera = new Camera();
-  }
-
-  dst.camera.setArmRotation(camPos.horizontal, camPos.vertical);
-  dst.camera.setArmLength(camPos.distance);
-};
-
-
-Graph3d.prototype._setStyle = function (style, dst) {
-  if (style === undefined) {
-    return;   // Nothing to do
-  }
-
-  var styleNumber;
-
-  if (typeof style === 'string') {
-    styleNumber = StyleHandler.getStyleNumber(style);
-
-   if (styleNumber === -1 ) {
-     return;   // string value of style invalid. TODO: generate error here?
-   }
-  } else {
-    // Do a pedantic check on style number value
-    var valid = false;
-    for (var n in StyleHandler.STYLE) {
-      if (StyleHandler.STYLE[n] === style) {
-        valid = true;
-        break;
-      }
-    }
-
-    if (!valid) {
-      throw 'Value style = ' + style + ' is invalid';
-    }
-
-    styleNumber = style;
-  }
-
-  dst.style         = styleNumber;
-  dst._styleHandler = StyleHandler.init(dst.style);
-};
-
-
-/**
- * Merge the values from src into dst
- *
- * src values are considered to be user-defined options.
- * dst values are the currently active values in the Graph3d instance.
- */
-Graph3d.prototype._mergeOptions = function (src, dst) {
-  if (src === undefined) {
-    return;   // Nothing to do
-  }
-
-  // retrieve simple parameter values
-  support.safeCopy(src, dst, OPTIONKEYS);
-
-  //
-  // Special handling for certain parameters
-  //
-  this._setStyle(src.style, dst);
-
-  if (src.tooltip !== undefined) {
-    dst.showTooltip       = src.tooltip;
-  }
-  if (src.backgroundColor !== undefined) {
-    dst._setBackgroundColor(src.backgroundColor);
-  }
-
-  this._setCameraPosition(src.cameraPosition, dst);
-  this._setDataColor(src.dataColor, dst);
-};
-
-
 /**
  * Update the options. Options will be merged with current options
  * @param {Object} options
  */
 Graph3d.prototype.setOptions = function (options) {
   this.animationStop();
-  this._mergeOptions(options, this);
+  Settings.merge(options, this);
   this.setSize(this.width, this.height);
 
   // re-load the data
@@ -922,9 +621,9 @@ Graph3d.prototype._redrawLegend = function() {
   //Return without drawing anything, if no legend is specified
   if (this.showLegend !== true) return;
 
-  var ctx = this.getContext();
+  var ctx       = this.getContext();
   ctx.lineWidth = 1;
-  ctx.font = '14px arial'; // TODO: put in options
+  ctx.font      = '14px arial'; // TODO: put in options
 
   this._styleHandler.drawLegend(this, ctx);
 };
