+++ lib/graph3d/stylehandlers/BarSizeStyleHandler.js
@@ -29,11 +29,12 @@ BarSizeStyleHandler.prototype.getColor = function(graph3d, point) {
 BarSizeStyleHandler.prototype.redrawPoint =
 function(graph3d, ctx, point, xWidth, yWidth) {
   // calculate size for the bar
-  var minOffset = (point.point.value - graph3d.valueMin);
-  var denom     = (graph3d.valueMax  - graph3d.valueMin) * 0.8 + 0.2;
+  var numer  = (point.point.value - graph3d.valueMin);
+  var denom  = (graph3d.valueMax  - graph3d.valueMin);
+  var factor = numer/denom * 0.8 + 0.2;
 
-  xWidth = (graph3d.xBarWidth / 2) * minOffset / denom;
-  yWidth = (graph3d.yBarWidth / 2) * minOffset / denom;
+  xWidth = graph3d.xBarWidth / 2 * factor;
+  yWidth = graph3d.yBarWidth / 2 * factor;
 
   this.parent().redrawPoint.call(this, graph3d, ctx, point, xWidth, yWidth);
 };
