+++ lib/graph3d/stylehandlers/Support.js
@@ -8,6 +8,20 @@ function derive(child, parent) {
 
 
 /**
+ * Helper function to tone down the length of the construct:
+ *
+ *         target  = (A !== undefined) ? A : B;
+ */
+function select(valueA, valueB) {
+  if (valueA !== undefined) {
+    return valueA;
+  } else {
+    return valueB;
+  }
+}
+
+
+/**
  * Copy all fields from src to dst
  *
  * If a value in src is undefined, so be it.
@@ -79,6 +93,7 @@ function calcTranslations(graph3d, points, sort) {
 
 
 module.exports.derive              = derive;
+module.exports.select              = select;
 module.exports.copy                = copy;
 module.exports.safeCopy            = safeCopy;
 module.exports.calcTranslations    = calcTranslations;
